pyproject.toml
[project]
name = "contramate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "MrDataPsycho", email = "mr.data.psycho@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "opensearch-py>=2.4.0",
    "psycopg2-binary>=2.9.7",
    "boto3>=1.29.0",
    "sqlmodel>=0.0.14",
    "loguru>=0.7.3",
    "pydantic-settings>=2.10.1",
    "openai>=1.108.1",
    "azure-identity>=1.25.0",
    "pymupdf4llm>=0.0.27",
    "aioboto3>=15.1.0",
    "neopipe>=0.1.1",
    "tenacity>=9.1.2",
    "polars>=1.33.1",
]

[project.scripts]
contramate = "contramate.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
    "polars>=1.33.1",
    "rich>=14.1.0",
    "ruff>=0.14.0",
    "typer>=0.19.1",
]
